feat 11: Implement StackNavigator for pages
Multiple pages can be navigated with the implementation of a stack based
navigator function.
