+++ package.json
@@ -10,6 +10,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/stack": "^7.1.0",
     "expo": "~52.0.19",
     "expo-status-bar": "~2.0.0",
     "react": "18.3.1",
